🚚 rename specification to make it generic
