+++ domain/src/specifications/aggregate_exists.rs
